+++ src/main/java/insane96mcp/survivalreimagined/module/mining/feature/Keego.java
@@ -74,10 +74,10 @@ public class Keego extends Feature {
 	public static final RegistryObject<Item> HOE = SRItems.REGISTRY.register("keego_hoe", () -> new HoeItem(ITEM_TIER, -2, -1.0F, new Item.Properties()));
 
 	private static final SRArmorMaterial ARMOR_MATERIAL = new SRArmorMaterial("survivalreimagined:keego", 22, Util.make(new EnumMap<>(ArmorItem.Type.class), (p_266652_) -> {
-		p_266652_.put(ArmorItem.Type.BOOTS, 3);
+		p_266652_.put(ArmorItem.Type.BOOTS, 4);
 		p_266652_.put(ArmorItem.Type.LEGGINGS, 5);
 		p_266652_.put(ArmorItem.Type.CHESTPLATE, 6);
-		p_266652_.put(ArmorItem.Type.HELMET, 2);
+		p_266652_.put(ArmorItem.Type.HELMET, 3);
 	}), 6, SoundEvents.ARMOR_EQUIP_IRON, 1f, 0.03f, () -> Ingredient.of(GEM.get()));
 
 	public static final RegistryObject<Item> HELMET = SRItems.REGISTRY.register("keego_helmet", () -> new ArmorItem(ARMOR_MATERIAL, ArmorItem.Type.HELMET, new Item.Properties()));